requirements.txt
Django==4.2.0
psycopg2-binary==2.9.5
uvicorn==0.20.0